refactor: create custom hooks
